adda

+++ data_loader.py
@@ -38,7 +38,7 @@ class BERTDataset(Dataset):
         batch_label = []
         batch_idx = []
 
-        batch_label = torch.tensor([item[0]['relation'] for item in data]).float()
+        batch_label = torch.tensor([item[0]['relation'] for item in data])
         # batch_instance['ids'] = torch.tensor([item[0]['ids'] for item in data])
         # batch_instance['mask'] = torch.tensor([item[0]['mask'] for item in data])
         batch_idx = torch.tensor([item[1] for item in data])
